requirements.txt
Django==5.0
django-webtest>=1.9.11

gunicorn==22.0.0

dj-database-url==2.2.0
psycopg2==2.9.9

celery==5.4.0
django-picklefield==3.2

django-anymail==11.0.1

python-dateutil==2.9.0
unicodecsv==0.14.1
bleach==6.1.0
validate_email==1.3
pycryptodome==3.20.0

python3-openid==3.2.0
boto==2.49.0
django-ses==4.1.0
oauth2client==4.1.3
rollbar==1.0.0

#ldap auth
django-auth-ldap==4.8.0
python-ldap==3.4.4

requests
httplib2~=0.22.0